Added secure login and register
The project now features a secure login and registration function, it integrates with an sqlite3 database for users and provides a new session cookie with 15 minutes of age. I also added the Dockerfile and the docker-compose.yml, since they are auto-generated i might need to check them better, but that will be done later on.